修改电压数据类型为小数，取一位小数

+++ app01/models.py
@@ -26,7 +26,7 @@ class Raw_data(models.Model):
 	time = models.DateTimeField()
 	pressure = models.IntegerField()
 	temperature = models.IntegerField()
-	s_volt = models.IntegerField()
+	s_volt = models.FloatField()
 	s_press = models.IntegerField()
 	s_temp = models.IntegerField()
 	s_data = models.CharField(max_length=30)
